Move request_method to pusher.request module
